create restartJasper component
